rename callbacks in/out -> update/render
