+++ callbacks/render_combined_visualization.py
