+++ callbacks/render_search_results.py
